Organize event APIs
